drop support for ruby 2.6
