+++ .github/workflows/main.yml
@@ -31,7 +31,6 @@ jobs:
           - 3.1
           - 3.0
           - 2.7
-          - 2.6
 
     steps:
       - uses: actions/checkout@v3
