+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.14.0
 
+- *Breaking Change* - Support for Ruby 2.6 is dropped
 - Adjusts the migration and install generators to use the `fx` gem so that Rails 7+ can use `schema.rb`
   instead of `structure.sql`
