+++ hoardable.gemspec
@@ -12,7 +12,7 @@ Gem::Specification.new do |spec|
   spec.description = 'Rails model versioning with the power of uni-temporal inherited tables'
   spec.homepage = 'https://github.com/waymondo/hoardable'
   spec.license = 'MIT'
-  spec.required_ruby_version = '>= 2.6.0'
+  spec.required_ruby_version = '>= 2.7.0'
 
   spec.metadata['homepage_uri'] = spec.homepage
   spec.metadata['source_code_uri'] = spec.homepage
@@ -31,6 +31,6 @@ Gem::Specification.new do |spec|
   spec.add_dependency 'activesupport', '>= 6.1', '< 8'
   spec.add_dependency 'railties', '>= 6.1', '< 8'
 
-  spec.add_dependency 'fx', '>= 0', '< 1'
+  spec.add_dependency 'fx', '>= 0.8', '< 1'
   spec.add_dependency 'pg', '>= 1', '< 2'
 end
