+++ test/test_helper.rb
@@ -22,16 +22,7 @@ require 'active_model/railtie'
 require 'active_record/railtie'
 require 'action_text/engine'
 
-class Dummy < Rails::Application
-  config.load_defaults Rails::VERSION::STRING.to_f
-  config.eager_load = false
-  config.active_storage.service_configurations = {}
-  config.paths['config/database'] = ['test/config/database.yml']
-  config.paths['db/migrate'] = ['tmp/db/migrate'] if Hoardable::SUPPORTS_VIRTUAL_COLUMNS
-  config.active_record.encryption&.key_derivation_salt = SecureRandom.hex
-  config.active_record.encryption&.primary_key = SecureRandom.hex
-end
-
+require 'config/application'
 Rails.initialize!
 
 SUPPORTS_ENCRYPTED_ACTION_TEXT = ActiveRecord.version >= Gem::Version.new('7.0.4')
